rf: move `gen_row()` to `md.py`
