+++ src/dandisets_linkml_status_tools/tools/md.py
@@ -1,0 +1,13 @@
+# This file contains helpers for generating Markdown files
+from collections.abc import Iterable
+
+
+def gen_row(cell_str_values: Iterable[str]) -> str:
+    """
+    Construct a row of a Markdown table with given cell string values
+    :param cell_str_values: The given iterable of cell string values
+    :return: The constructed row of a Markdown table
+
+    Note: The given iterable of cell string values are `str` values
+    """
+    return f'|{"|".join(cell_str_values)}|\n'
